(spline) Combine spline src files
